refactor: consolidate projection utilities and apply Rust idioms
Move generic projection code to shared customscan/projections.rs.
Apply review feedback: Vec over HashMap, early-continue, .any(), iter_ptr().
