+++ pg_search/src/postgres/customscan/aggregatescan/mod.rs
@@ -47,6 +47,7 @@ use crate::postgres::customscan::builders::custom_state::{
     CustomScanStateBuilder, CustomScanStateWrapper,
 };
 use crate::postgres::customscan::explainer::Explainer;
+use crate::postgres::customscan::projections::{create_placeholder_targetlist, placeholder_procid};
 use crate::postgres::customscan::solve_expr::SolvePostgresExpressions;
 use crate::postgres::customscan::{
     range_table, CreateUpperPathsHookArgs, CustomScan, ExecMethod, PlainExecCapable,
@@ -334,55 +335,64 @@ impl CustomScan for AggregateScan {
                 // This matches pdbscan's approach of setting Const values directly.
                 let mut agg_iter = row.aggregates.iter();
                 for (i, entry) in state.custom_state().aggregate_clause.entries().enumerate() {
-                    if let TargetListEntry::Aggregate(agg_type) = entry {
-                        if let Some(&const_node) = state.custom_state().const_agg_nodes.get(&i) {
-                            // Get the next aggregate result
-                            let agg_result = agg_iter.next().and_then(|v| v.clone());
+                    let TargetListEntry::Aggregate(agg_type) = entry else {
+                        continue;
+                    };
 
-                            // Convert to datum using the Const node's type (native aggregate type)
-                            // not the output tuple descriptor's type
-                            let (datum, is_null) = if row.is_empty() {
-                                // Empty result - use nullish value
-                                let nullish_datum = agg_type.nullish().value.and_then(|value| {
-                                    TantivyValue(OwnedValue::F64(value))
-                                        .try_into_datum((*const_node).consttype.into())
-                                        .unwrap()
-                                });
-                                (
-                                    nullish_datum.unwrap_or(pg_sys::Datum::null()),
-                                    nullish_datum.is_none(),
-                                )
-                            } else if agg_type.can_use_doc_count()
-                                && !state.custom_state().aggregate_clause.has_filter()
-                                && state.custom_state().aggregate_clause.has_groupby()
-                            {
-                                let d = row
-                                    .doc_count()
-                                    .try_into_datum(pgrx::PgOid::from((*const_node).consttype));
-                                match d {
-                                    Ok(Some(datum)) => (datum, false),
-                                    _ => (pg_sys::Datum::null(), true),
-                                }
-                            } else {
-                                // Use the native aggregate result type (from the Const node)
-                                let d = exec::aggregate_result_to_datum(
-                                    agg_result,
-                                    agg_type,
-                                    (*const_node).consttype, // Use Const's type, not output type
-                                );
-                                match d {
-                                    Some(datum) => (datum, false),
-                                    None => (pg_sys::Datum::null(), true),
-                                }
-                            };
+                    let Some(const_node) = state
+                        .custom_state()
+                        .const_agg_nodes
+                        .get(i)
+                        .copied()
+                        .flatten()
+                    else {
+                        // No Const node for this aggregate, skip the iterator
+                        agg_iter.next();
+                        continue;
+                    };
 
-                            (*const_node).constvalue = datum;
-                            (*const_node).constisnull = is_null;
-                        } else {
-                            // No Const node for this aggregate, skip the iterator
-                            agg_iter.next();
+                    // Get the next aggregate result
+                    let agg_result = agg_iter.next().and_then(|v| v.clone());
+
+                    // Convert to datum using the Const node's type (native aggregate type)
+                    // not the output tuple descriptor's type
+                    let (datum, is_null) = if row.is_empty() {
+                        // Empty result - use nullish value
+                        let nullish_datum = agg_type.nullish().value.and_then(|value| {
+                            TantivyValue(OwnedValue::F64(value))
+                                .try_into_datum((*const_node).consttype.into())
+                                .unwrap()
+                        });
+                        (
+                            nullish_datum.unwrap_or(pg_sys::Datum::null()),
+                            nullish_datum.is_none(),
+                        )
+                    } else if agg_type.can_use_doc_count()
+                        && !state.custom_state().aggregate_clause.has_filter()
+                        && state.custom_state().aggregate_clause.has_groupby()
+                    {
+                        let d = row
+                            .doc_count()
+                            .try_into_datum(pgrx::PgOid::from((*const_node).consttype));
+                        match d {
+                            Ok(Some(datum)) => (datum, false),
+                            _ => (pg_sys::Datum::null(), true),
                         }
-                    }
+                    } else {
+                        // Use the native aggregate result type (from the Const node)
+                        let d = exec::aggregate_result_to_datum(
+                            agg_result,
+                            agg_type,
+                            (*const_node).consttype, // Use Const's type, not output type
+                        );
+                        match d {
+                            Some(datum) => (datum, false),
+                            None => (pg_sys::Datum::null(), true),
+                        }
+                    };
+
+                    (*const_node).constvalue = datum;
+                    (*const_node).constisnull = is_null;
                 }
 
                 // Set the scan tuple for expression evaluation context
@@ -500,21 +510,14 @@ unsafe fn replace_aggrefs_in_target_list(plan: *mut pg_sys::Plan) {
         }
     }
 
-    let targetlist = (*plan).targetlist;
-    let list_len = pg_sys::list_length(targetlist);
+    let targetlist = PgList::<pg_sys::TargetEntry>::from_pg((*plan).targetlist);
 
     // Check if there are any Aggref nodes anywhere in the target list
-    let mut has_aggref = false;
-    for i in 0..list_len {
-        let te = pg_sys::list_nth(targetlist, i) as *mut pg_sys::TargetEntry;
-        if !te.is_null() && !(*te).expr.is_null() {
-            // Use expression_tree_walker to check for nested Aggrefs
-            if expr_contains_aggref((*te).expr as *mut pg_sys::Node) {
-                has_aggref = true;
-                break;
-            }
-        }
-    }
+    let has_aggref = targetlist.iter_ptr().any(|te| {
+        !te.is_null()
+            && !(*te).expr.is_null()
+            && expr_contains_aggref((*te).expr as *mut pg_sys::Node)
+    });
 
     if !has_aggref {
         return;
@@ -522,8 +525,7 @@ unsafe fn replace_aggrefs_in_target_list(plan: *mut pg_sys::Plan) {
 
     // Build a new target list with Aggrefs replaced by placeholders
     let mut new_targetlist: *mut pg_sys::List = std::ptr::null_mut();
-    for i in 0..list_len {
-        let te = pg_sys::list_nth(targetlist, i) as *mut pg_sys::TargetEntry;
+    for te in targetlist.iter_ptr() {
         let new_te = pg_sys::flatCopyTargetEntry(te);
 
         // Use the mutator to replace any Aggref nodes in the expression
@@ -534,233 +536,6 @@ unsafe fn replace_aggrefs_in_target_list(plan: *mut pg_sys::Plan) {
     }
 
     (*plan).targetlist = new_targetlist;
-}
-
-/// Context for the Const placeholder mutator
-struct ConstPlaceholderContext {
-    /// Current target entry index being processed
-    current_te_idx: usize,
-    /// The OID of our placeholder function (pdb.agg_fn)
-    placeholder_funcid: pg_sys::Oid,
-    /// Collected Const node pointers for later mutation
-    /// Key: target entry index, Value: pointer to Const node
-    const_nodes: crate::api::HashMap<usize, *mut pg_sys::Const>,
-}
-
-/// Create a placeholder target list with Const nodes replacing our FuncExpr placeholders.
-/// This is called AFTER replace_aggrefs_in_target_list has replaced Aggrefs with FuncExprs.
-/// For wrapped expressions, we replace those FuncExprs with Const nodes that will be
-/// mutated with actual aggregate values before each ExecBuildProjectionInfo call.
-/// This follows the pdbscan pattern where Const values are baked in when projection is built.
-///
-/// Returns: (placeholder_targetlist, const_nodes, needs_projection)
-/// - placeholder_targetlist: target list with FuncExprs replaced by Const nodes
-/// - const_nodes: HashMap of Const node pointers for later mutation
-/// - needs_projection: true if projection is needed (wrapped expressions exist)
-unsafe fn create_placeholder_targetlist(
-    targetlist: *mut pg_sys::List,
-) -> (
-    *mut pg_sys::List,
-    crate::api::HashMap<usize, *mut pg_sys::Const>,
-    bool,
-) {
-    use pgrx::pg_guard;
-
-    if targetlist.is_null() {
-        return (std::ptr::null_mut(), Default::default(), false);
-    }
-
-    let placeholder_funcid = placeholder_procid();
-    let list_len = pg_sys::list_length(targetlist);
-
-    // Check if any target entries have wrapped placeholder FuncExprs (not top-level)
-    let mut needs_projection = false;
-    for i in 0..list_len {
-        let te = pg_sys::list_nth(targetlist, i) as *mut pg_sys::TargetEntry;
-        if !te.is_null() && !(*te).expr.is_null() {
-            // Check if the expression is NOT a direct FuncExpr placeholder but CONTAINS one
-            let is_top_level_placeholder = (*(*te).expr).type_ == pg_sys::NodeTag::T_FuncExpr
-                && (*((*te).expr as *mut pg_sys::FuncExpr)).funcid == placeholder_funcid;
-
-            if !is_top_level_placeholder
-                && expr_contains_placeholder_funcexpr(
-                    (*te).expr as *mut pg_sys::Node,
-                    placeholder_funcid,
-                )
-            {
-                needs_projection = true;
-                break;
-            }
-        }
-    }
-
-    if !needs_projection {
-        return (std::ptr::null_mut(), Default::default(), false);
-    }
-
-    #[pg_guard]
-    unsafe extern "C-unwind" fn const_placeholder_mutator(
-        node: *mut pg_sys::Node,
-        context: *mut core::ffi::c_void,
-    ) -> *mut pg_sys::Node {
-        if node.is_null() {
-            return std::ptr::null_mut();
-        }
-
-        let ctx = &mut *(context as *mut ConstPlaceholderContext);
-
-        // If this is our placeholder FuncExpr, replace it with a Const
-        if (*node).type_ == pg_sys::NodeTag::T_FuncExpr {
-            let funcexpr = node as *mut pg_sys::FuncExpr;
-            if (*funcexpr).funcid == ctx.placeholder_funcid {
-                // Create a Const node with NULL value (will be mutated before projection)
-                let const_node = make_placeholder_const_from_funcexpr(funcexpr);
-                // Store the pointer for later mutation
-                debug_assert!(
-                    !ctx.const_nodes.contains_key(&ctx.current_te_idx),
-                    "AggregateScan supports only one aggregate per target entry"
-                );
-                ctx.const_nodes.insert(ctx.current_te_idx, const_node);
-                return const_node as *mut pg_sys::Node;
-            }
-        }
-
-        // For all other nodes, use the standard mutator to walk children
-        #[cfg(not(any(feature = "pg16", feature = "pg17", feature = "pg18")))]
-        {
-            let fnptr = const_placeholder_mutator as usize as *const ();
-            let mutator: unsafe extern "C-unwind" fn() -> *mut pg_sys::Node =
-                std::mem::transmute(fnptr);
-            pg_sys::expression_tree_mutator(node, Some(mutator), context)
-        }
-
-        #[cfg(any(feature = "pg16", feature = "pg17", feature = "pg18"))]
-        {
-            pg_sys::expression_tree_mutator_impl(node, Some(const_placeholder_mutator), context)
-        }
-    }
-
-    let mut ctx = ConstPlaceholderContext {
-        current_te_idx: 0,
-        placeholder_funcid,
-        const_nodes: Default::default(),
-    };
-
-    // Build a new target list with ALL FuncExpr placeholders replaced by Const nodes.
-    // This is critical for mixed wrapped/unwrapped cases like:
-    //   SELECT pdb.agg(...), (pdb.agg(...))->'avg' FROM ...
-    // If we only replace wrapped ones, ExecProject will try to execute the unwrapped
-    // pdb.agg_fn() FuncExpr and fail with "placeholder should not be executed".
-    let mut new_targetlist: *mut pg_sys::List = std::ptr::null_mut();
-    for i in 0..list_len {
-        let te = pg_sys::list_nth(targetlist, i) as *mut pg_sys::TargetEntry;
-        let new_te = pg_sys::flatCopyTargetEntry(te);
-
-        ctx.current_te_idx = i as usize;
-
-        // AggregateScan assumes a contiguous, non-resjunk targetlist.
-        // Wrapper support doesn't change this assumption.
-        debug_assert!(
-            !(*te).resjunk,
-            "AggregateScan does not support resjunk target entries"
-        );
-        debug_assert!(
-            (*te).resno as usize == (i as usize) + 1,
-            "AggregateScan expects contiguous resno values (1, 2, 3, ...)"
-        );
-
-        // Check if this expression contains any placeholder FuncExpr (wrapped or top-level)
-        let is_top_level_placeholder = (*(*te).expr).type_ == pg_sys::NodeTag::T_FuncExpr
-            && (*((*te).expr as *mut pg_sys::FuncExpr)).funcid == placeholder_funcid;
-
-        let contains_placeholder = is_top_level_placeholder
-            || expr_contains_placeholder_funcexpr(
-                (*te).expr as *mut pg_sys::Node,
-                placeholder_funcid,
-            );
-
-        if contains_placeholder {
-            // Replace ALL placeholder FuncExprs with Const nodes (both wrapped and top-level)
-            // For top-level: the mutator will replace the FuncExpr directly with a Const
-            // For wrapped: the mutator will walk the tree and replace nested FuncExprs
-            let ctx_ptr = &mut ctx as *mut ConstPlaceholderContext as *mut core::ffi::c_void;
-            let new_expr = const_placeholder_mutator((*te).expr as *mut pg_sys::Node, ctx_ptr);
-            (*new_te).expr = new_expr as *mut pg_sys::Expr;
-        }
-
-        new_targetlist = pg_sys::lappend(new_targetlist, new_te.cast());
-    }
-
-    (new_targetlist, ctx.const_nodes, true)
-}
-
-/// Check if an expression tree contains our placeholder FuncExpr
-unsafe fn expr_contains_placeholder_funcexpr(
-    node: *mut pg_sys::Node,
-    placeholder_funcid: pg_sys::Oid,
-) -> bool {
-    use pgrx::pg_guard;
-    use std::ptr::addr_of_mut;
-
-    struct WalkerContext {
-        found: bool,
-        funcid: pg_sys::Oid,
-    }
-
-    #[pg_guard]
-    unsafe extern "C-unwind" fn walker(
-        node: *mut pg_sys::Node,
-        context: *mut core::ffi::c_void,
-    ) -> bool {
-        if node.is_null() {
-            return false;
-        }
-
-        let ctx = &mut *(context as *mut WalkerContext);
-
-        if (*node).type_ == pg_sys::NodeTag::T_FuncExpr {
-            let funcexpr = node as *mut pg_sys::FuncExpr;
-            if (*funcexpr).funcid == ctx.funcid {
-                ctx.found = true;
-                return true; // Stop walking
-            }
-        }
-
-        pg_sys::expression_tree_walker(node, Some(walker), context)
-    }
-
-    let mut ctx = WalkerContext {
-        found: false,
-        funcid: placeholder_funcid,
-    };
-    walker(node, addr_of_mut!(ctx).cast());
-    ctx.found
-}
-
-/// Create a placeholder Const node from a FuncExpr placeholder.
-/// The Const will be initialized with NULL value and will be mutated with actual
-/// aggregate values before each ExecBuildProjectionInfo call. This follows the
-/// pdbscan pattern where Const values are baked in when projection is built per-row.
-unsafe fn make_placeholder_const_from_funcexpr(
-    funcexpr: *mut pg_sys::FuncExpr,
-) -> *mut pg_sys::Const {
-    let result_type = (*funcexpr).funcresulttype;
-
-    // Get type information for the Const node
-    let mut typlen: i16 = 0;
-    let mut typbyval: bool = false;
-    pg_sys::get_typlenbyval(result_type, &mut typlen, &mut typbyval);
-
-    // Create a Const with NULL value - will be mutated before each projection build
-    pg_sys::makeConst(
-        result_type,
-        -1,                     // typmod
-        (*funcexpr).funccollid, // collation
-        typlen as i32,          // constlen
-        pg_sys::Datum::null(),  // constvalue (NULL initially)
-        true,                   // constisnull (starts as NULL)
-        typbyval,               // constbyval
-    )
 }
 
 /// Check if an expression tree contains any Aggref nodes
@@ -863,16 +638,4 @@ unsafe fn make_text_const(text: &str) -> *mut pg_sys::Const {
         false, // constisnull
         false, // constbyval (text is not passed by value)
     )
-}
-
-/// Get the Oid of a placeholder function to use in the target list of aggregate custom scans.
-unsafe fn placeholder_procid() -> pg_sys::Oid {
-    let agg_fn_oid = crate::api::agg_fn_oid();
-    if agg_fn_oid != pg_sys::InvalidOid {
-        agg_fn_oid
-    } else {
-        // Fallback to now() if pdb.agg_fn doesn't exist yet (e.g., during extension creation)
-        pgrx::direct_function_call::<pg_sys::Oid>(pg_sys::regprocedurein, &[c"now()".into_datum()])
-            .expect("the `now()` function should exist")
-    }
 }
